FB-657 flow: fix tests immutablex: fix scanner issue, add apikey support, fix asset serialization

+++ integration/immutablex/src/main/kotlin/com/rarible/protocol/union/integration/immutablex/converter/ImmutablexItemConverter.kt
@@ -31,7 +31,7 @@ object ImmutablexItemConverter {
     private fun convertInternal(asset: ImmutablexAsset, blockchain: BlockchainDto): UnionItem {
         val deleted = asset.user!! == "${Address.ZERO()}"
         return UnionItem(
-            id = ItemIdDto(BlockchainDto.IMMUTABLEX, contract = asset.tokenAddress, tokenId = asset.tokenId.toBigInteger()),
+            id = ItemIdDto(BlockchainDto.IMMUTABLEX, contract = asset.tokenAddress, tokenId = asset.tokenId),
             collection = CollectionIdDto(blockchain, asset.tokenAddress),
             creators = emptyList(),
             owners = emptyList(),
